requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
biopython>=1.79
pyyaml>=5.4.1

# Machine learning (TIGER)
tensorflow>=2.11.0
# OR for CPU-only:
# tensorflow-cpu>=2.11.0

# TIGER dependencies
keras>=2.11.0

# Utilities
tqdm>=4.62.0
colorama>=0.4.4
requests>=2.26.0

# Optional
matplotlib>=3.4.0
seaborn>=0.11.0

# SHAP (for TIGER model interpretation - optional)
# shap>=0.41.0
